+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New #144: Add `chatId` parameter to `SendGift` method.
 - New #144: Add `canSendGift` field to `ChatFullInfo` type.
 - New #144: Add `TransactionPartnerChat` type.
+- New #144: Add `cover` and `startTimestamp` fields to `Video` type.
 
 ## 0.7.0 February 10, 2025